re-use functionality from color manipulation for 3d scatterplot
